Draft 1 Voter Registration Demographics
I've added the first draft of the North Carolina Voter Registration Demographics report for everyone's review.  This report is based upon a small subset of the data, so it will need to be updated once all the data is in.  I've added some analyses for each section as an example of how we can state conclusions.