Fixes Filetracker to use known name rather than querying...:
When two files were extracted at the same moment, they were going into
the same folder. This meant that querying for the resulting filename had
a subtle race condition. Essentially, any two files extracted into the
same folder would "collide" in a sense, and the data recorded into the
FileTracker model would match the first file found. Instead, the
approach taken here is to assume the resulting file is a .txt of
whatever the zip filename is. This means that collisions are irrelevant
to how the database tracks these filenames and should be more resilient.